Added First Modifier!
Added a modifier class. Anything that can be anchored to a cube must
share the "CubeModifier" tag. Added a Thruster modifier that propels a
cube. Added a selection manager that through number presses determines
if we are adding cubes, adding cubemodifiers, or activating triggers on
cubemodifiers. (1)Allows the placing of cubes (2) Allows placing of
thruster modifiers (3) Allows activation of triggers.